+++ test/solc/CommandLineParser.cpp
@@ -251,7 +251,6 @@ BOOST_AUTO_TEST_CASE(no_import_callback)
 	std::vector<std::vector<std::string>> commandLinePerInputMode = {
 		{"solc", "--no-import-callback", "contract.sol"},
 		{"solc", "--standard-json", "--no-import-callback", "input.json"},
-		{"solc", "--assemble", "--no-import-callback", "input.yul"},
 		{"solc", "--strict-assembly", "--no-import-callback", "input.yul"},
 		{"solc", "--experimental", "--import-ast", "--no-import-callback", "ast.json"},
 		{"solc", "--link", "--no-import-callback", "input.bin"},
@@ -274,11 +273,8 @@ BOOST_AUTO_TEST_CASE(via_ir_options)
 BOOST_AUTO_TEST_CASE(assembly_mode_options)
 {
 	static std::vector<std::tuple<std::vector<std::string>, YulStack::Machine, YulStack::Language>> const allowedCombinations = {
-		{{"--machine=evm", "--yul-dialect=evm", "--assemble"}, YulStack::Machine::EVM, YulStack::Language::StrictAssembly},
 		{{"--machine=evm", "--yul-dialect=evm", "--strict-assembly"}, YulStack::Machine::EVM, YulStack::Language::StrictAssembly},
-		{{"--machine=evm", "--assemble"}, YulStack::Machine::EVM, YulStack::Language::StrictAssembly},
 		{{"--machine=evm", "--strict-assembly"}, YulStack::Machine::EVM, YulStack::Language::StrictAssembly},
-		{{"--assemble"}, YulStack::Machine::EVM, YulStack::Language::StrictAssembly},
 		{{"--strict-assembly"}, YulStack::Machine::EVM, YulStack::Language::StrictAssembly},
 	};
 
@@ -429,20 +425,20 @@ BOOST_AUTO_TEST_CASE(invalid_options_input_modes_combinations)
 {
 	std::map<std::string, std::vector<std::string>> invalidOptionInputModeCombinations = {
 		// TODO: This should eventually contain all options.
-		{"--experimental-via-ir", {"--assemble", "--strict-assembly", "--standard-json", "--link"}},
-		{"--via-ir", {"--assemble", "--strict-assembly", "--standard-json", "--link"}},
-		{"--metadata-literal", {"--assemble", "--strict-assembly", "--standard-json", "--link"}},
-		{"--metadata-hash=swarm", {"--assemble", "--strict-assembly", "--standard-json", "--link"}},
-		{"--model-checker-show-proved-safe", {"--assemble", "--strict-assembly", "--standard-json", "--link"}},
-		{"--model-checker-show-unproved", {"--assemble", "--strict-assembly", "--standard-json", "--link"}},
-		{"--model-checker-show-unsupported", {"--assemble", "--strict-assembly", "--standard-json", "--link"}},
-		{"--model-checker-div-mod-no-slacks", {"--assemble", "--strict-assembly", "--standard-json", "--link"}},
-		{"--model-checker-engine=bmc", {"--assemble", "--strict-assembly", "--standard-json", "--link"}},
-		{"--model-checker-invariants=contract,reentrancy", {"--assemble", "--strict-assembly", "--standard-json", "--link"}},
-		{"--model-checker-solvers=z3,smtlib2", {"--assemble", "--strict-assembly", "--standard-json", "--link"}},
-		{"--model-checker-timeout=5", {"--assemble", "--strict-assembly", "--standard-json", "--link"}},
-		{"--model-checker-contracts=contract1.yul:A,contract2.yul:B", {"--assemble", "--strict-assembly", "--standard-json", "--link"}},
-		{"--model-checker-targets=underflow,divByZero", {"--assemble", "--strict-assembly", "--standard-json", "--link"}},
+		{"--experimental-via-ir", {"--strict-assembly", "--standard-json", "--link"}},
+		{"--via-ir", {"--strict-assembly", "--standard-json", "--link"}},
+		{"--metadata-literal", {"--strict-assembly", "--standard-json", "--link"}},
+		{"--metadata-hash=swarm", {"--strict-assembly", "--standard-json", "--link"}},
+		{"--model-checker-show-proved-safe", {"--strict-assembly", "--standard-json", "--link"}},
+		{"--model-checker-show-unproved", {"--strict-assembly", "--standard-json", "--link"}},
+		{"--model-checker-show-unsupported", {"--strict-assembly", "--standard-json", "--link"}},
+		{"--model-checker-div-mod-no-slacks", {"--strict-assembly", "--standard-json", "--link"}},
+		{"--model-checker-engine=bmc", {"--strict-assembly", "--standard-json", "--link"}},
+		{"--model-checker-invariants=contract,reentrancy", {"--strict-assembly", "--standard-json", "--link"}},
+		{"--model-checker-solvers=z3,smtlib2", {"--strict-assembly", "--standard-json", "--link"}},
+		{"--model-checker-timeout=5", {"--strict-assembly", "--standard-json", "--link"}},
+		{"--model-checker-contracts=contract1.yul:A,contract2.yul:B", {"--strict-assembly", "--standard-json", "--link"}},
+		{"--model-checker-targets=underflow,divByZero", {"--strict-assembly", "--standard-json", "--link"}},
 		{"--via-ssa-cfg", {"--standard-json", "--link", "--import-asm-json"}}
 	};